requirements.txt
flask==2.2.5
gunicorn==21.2.0
langchain_pinecone==0.0.3
langchain_openai==0.1.1
langchain_core==0.1.35
python-dotenv==1.0.1